refactor(calendar): インラインタグパレットを作成ロジックとジェスチャに分割する
InlineTagPalette.tsx（522行）から entry 作成・競合判定を
useInlineTagPaletteCreation hook へ、リサイズ / long-press 移動を
inline-selection-gestures の factory へ移動。hook は store を直接購読し、
cross-feature import は barrel 経由のまま増やしていない。

Refs #1527
